added captouch drivers
Basic captouch infrastructure in place. You can run 'tchraw' and
get a live reading of the electrode values from the captouch sensor
to confirm that basic functionality is there.

Had to change the I2C port to #2, because I forgot that we wired
it to I2C1 and not I2C0 (and that the interfaces start numbering
from 1, not 0).